// TRC-Beneficios-Back/src/interfaces/IProtocol.ts
export interface protocoloAttributes {
  n_protocolo?: number;
  servico: string;
  status: string;
  beneficiarios_cpf: number;
  tipo: string;
  data_analise: Date;
  data_docpen?: Date;
  data_producao?: Date;
  data_enviado?: Date;
  data_entregue?: Date;
  observacoes?: string;
  entrega_crd: number;
  
}